feat(display): Optimize memory management by using 32bit buffer for display

+++ main/peripherals/display/sh1106.h
@@ -14,7 +14,7 @@ typedef struct {
     int height;
     int width;
     int transmission_failures;
-    uint8_t **buffer;
+    uint32_t **buffer;
 } SH1106Config;
 
 typedef enum {
